requirements.txt
aiohttp
python-dotenv
feedparser